toast add

+++ client/src/pages/Books.tsx
@@ -8,6 +8,7 @@ import {
 } from "@/components/ui/table";
 import { useDeleteBookMutation, useGetBooksQuery } from "@/redux/api/baseApi";
 import { Edit, Gift, Trash } from "lucide-react";
+import Swal from "sweetalert2";
 
 export function Books() {
   const { data, isLoading } = useGetBooksQuery([]);
@@ -17,7 +18,26 @@ export function Books() {
   }
 
   const handleBookDelete = async (bookId: any) => {
-    const res = await deleteBook(bookId);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const res = await deleteBook(bookId);
+        if (res) {
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your book has been deleted.",
+            icon: "success",
+          });
+        }
+      }
+    });
   };
 
   return (
